Reworked the app.jsx file so it can render the home page and any other pages. Created Pages folder to house all the templates for each page ( Home page, list page )
