+++ task-manager/src/components/Header.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 
-function Header() {
+function Header(props) {
   return (
     <header>
-      <h1>Dashboard (homepage)</h1>
+      <h1>{props.title}</h1>
     </header>
   );
 }
