+++ task-manager/src/index.css
@@ -64,7 +64,6 @@ big_box {
   position: absolute;
   padding: 16px 32px;
   box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.3);
-  background-color: ;
   text-align: center;
   width: 95%;
   height: 90%;
